+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Storage System
 - Storage Volume
 - Storage Volume Attachment
+- Storage Volume Template
 
 # [v1.2.0]
 ### Notes